Feature #1236: Avoid receiver-side copy for large contiguous messages
