+++ src/ck-core/charm++.h
@@ -173,93 +173,8 @@ public:
 
 #include "ckrdma.h"
 
-/********************* Superclass of all Chares ******************/
-#if CMK_MULTIPLE_DELETE
-#define CHARM_INPLACE_NEW \
-    void *operator new(size_t, void *ptr) { return ptr; }; \
-    void operator delete(void*, void*) {}; \
-    void *operator new(size_t s) { return malloc(s); } \
-    void operator delete(void *ptr) { free(ptr); }
-#else
-#define CHARM_INPLACE_NEW \
-    void *operator new(size_t, void *ptr) { return ptr; }; \
-    void *operator new(size_t s) { return malloc(s); } \
-    void operator delete(void *ptr) { free(ptr); }
-#endif
-
 // for object message queue
 #include "ckobjQ.h"
-#if CMK_SMP && CMK_TASKQUEUE
-#include "cktaskQ.h"
-#endif
-#if (defined(_FAULT_MLOG_) || defined(_FAULT_CAUSAL_))
-class ChareMlogData;
-#endif
-
-#define CHARE_MAGIC    0x201201
-
-/**
-  The base class of all parallel objects in Charm++,
-  including Array Elements, Groups, and NodeGroups.
-*/
-class Chare {
-  protected:
-    CkChareID thishandle;
-#if CMK_OBJECT_QUEUE_AVAILABLE
-    CkObjectMsgQ objQ;                // object message queue
-#endif
-  public:
-#if CMK_ERROR_CHECKING
-    int magic;
-#endif
-#ifndef CMK_CHARE_USE_PTR
-    int chareIdx;                  // index in the chare obj table (chare_objs)
-#endif
-#if (defined(_FAULT_MLOG_) || defined(_FAULT_CAUSAL_))
-    ChareMlogData *mlogData;
-#endif
-    Chare(CkMigrateMessage *m);
-    Chare();
-    virtual ~Chare(); //<- needed for *any* child to have a virtual destructor
-
-    /// Pack/UnPack - tell the runtime how to serialize this class's
-    /// data for migration, checkpoint, etc.
-    virtual void pup(PUP::er &p);
-    /// Routine that runtime the runtime actually calls, to enable
-    /// more intelligence in generated code that overrides this. The
-    /// actual pup() method must remain virtual, so that this
-    /// continues to work for older code.
-    virtual void virtual_pup(PUP::er &p) { pup(p); }
-    void parent_pup(PUP::er &p) {
-      (void)p;
-      CkAbort("Should never get here - only called in generated CBase code");
-    }
-
-    inline const CkChareID &ckGetChareID(void) const {return thishandle;}
-    inline void CkGetChareID(CkChareID *dest) const {*dest=thishandle;}
-    // object message queue
-    void  CkEnableObjQ();
-#if CMK_OBJECT_QUEUE_AVAILABLE
-    inline CkObjectMsgQ &CkGetObjQueue() { return objQ; }
-#endif
-    CHARM_INPLACE_NEW
-    /// Return the type of this chare, as present in _chareTable
-    virtual int ckGetChareType() const;
-    /// Return a strdup'd array containing this object's string name.
-    virtual char *ckDebugChareName(void);
-    /// Place into str a copy of the id of this object up to limit bytes, return
-    /// the number of bytes used for the id
-    virtual int ckDebugChareID(char *str, int limit);
-    virtual void ckDebugPup(PUP::er &p);
-    /// Called when a [threaded] charm entry method is created:
-    virtual void CkAddThreadListeners(CthThread tid, void *msg);
-#if CMK_ERROR_CHECKING
-    inline void sanitycheck() { 
-        if (magic != CHARE_MAGIC)
-          CmiAbort("Charm++ Fatal Error> Chare magic number does not agree, possibly due to pup functions not calling parent class.");
-    }
-#endif
-};
 
 #if CMK_HAS_IS_CONSTRUCTIBLE
 #include <type_traits>
@@ -306,39 +221,11 @@ struct call_migration_constructor
 #endif
 
 
-//Superclass of all Groups that cannot participate in reductions.
-//  Undocumented: should only be used inside Charm++.
+#include "Chare.h"
+
 /*forward*/ class Group;
-class IrrGroup : public Chare {
-  protected:
-    CkGroupID thisgroup;
-  public:
-    IrrGroup(CkMigrateMessage *m): Chare(m) { }
-    IrrGroup();
-    virtual ~IrrGroup(); //<- needed for *any* child to have a virtual destructor
 
-    virtual void pup(PUP::er &p);//<- pack/unpack routine
-    virtual void ckJustMigrated(void);
-    inline const CkGroupID &ckGetGroupID(void) const {return thisgroup;}
-    inline CkGroupID CkGetGroupID(void) const {return thisgroup;}
-    virtual int ckGetChareType() const;
-    virtual char *ckDebugChareName();
-    virtual int ckDebugChareID(char *, int);
-
-    // Silly run-time type information
-    virtual bool isNodeGroup() { return false; };
-    virtual bool isLocMgr(void){ return false; }
-    virtual bool isArrMgr(void){ return false; }
-    virtual bool isReductionMgr(void){ return false; }
-    static bool isIrreducible(){ return true;}
-    virtual void flushStates() {}
-		/*
-			FAULT_EVAC
-		*/
-		virtual void evacuate(){};
-		virtual void doneEvacuate(){};
-    virtual void CkAddThreadListeners(CthThread tid, void *msg);
-};
+#include "IrrGroup.h"
 
 // As described in http://www.gotw.ca/publications/mxc++-item-4.htm
 template<typename D, typename B>
@@ -392,6 +279,11 @@ template <typename T>
 void recursive_pup(T *obj, PUP::er &p) {
   recursive_pup_impl<T, IsDerivedFrom<T, CBase>::Is>()(obj, p);
 }
+
+
+
+//#include "ckreduction.h"
+
 
 // CBaseX::pup must be an empty override, so that the recursive PUPing
 // doesn't call an implementation multiple times up the inheritance
@@ -1041,8 +933,12 @@ typedef CProxySection_Group CProxySection_IrrGroup;
 
 //(CProxy_ArrayBase is defined in ckarray.h)
 
+#include "CkGroupInitCallback.h"
+
 //Defines the actual "Group"
 #include "ckreduction.h"
+
+
 
 class CkQdMsg {
   public:
