+++ src/ck-ldb/DistributedLB.ci
@@ -1,6 +1,8 @@
 module DistributedLB {
 
 extern module DistBaseLB;
+include "DistBaseLB.h";
+
 initnode void lbinit(void);
 
 group [migratable] DistributedLB : DistBaseLB {
