+++ src/ck-ldb/DummyLB.ci
@@ -1,6 +1,8 @@
 module DummyLB {
 
 extern module CentralLB;
+include "CentralLB.h";
+
 initnode void lbinit(void);
 group [migratable] DummyLB : CentralLB {
   entry void DummyLB(const CkLBOptions &);
