+++ src/ck-ldb/GreedyLB.ci
@@ -1,6 +1,8 @@
 module GreedyLB {
 
 extern module CentralLB;
+include "CentralLB.h";
+
 initnode void lbinit(void);
 group [migratable] GreedyLB : CentralLB {
   entry void GreedyLB(const CkLBOptions &);
