+++ src/ck-ldb/NodeLevelLB.ci
@@ -2,6 +2,7 @@ module NodeLevelLB {
 
 extern module LBDatabase;
 extern module HybridBaseLB;
+include "HybridBaseLB.h";
 
 initnode void lbinit(void);
