+++ src/ck-ldb/RandCentLB.ci
@@ -1,6 +1,7 @@
 module RandCentLB {
 
 extern module CentralLB;
+include "CentralLB.h";
 initnode void lbinit(void);
 group [migratable] RandCentLB : CentralLB {
   entry void RandCentLB(const CkLBOptions &);
