+++ src/ck-ldb/RefineLB.ci
@@ -1,6 +1,7 @@
 module RefineLB {
 
 extern module CentralLB;
+include "CentralLB.h";
 initnode void lbinit(void);
 group [migratable] RefineLB : CentralLB {
   entry void RefineLB(const CkLBOptions &);
