+++ src/ck-ldb/RotateLB.ci
@@ -2,6 +2,8 @@ module RotateLB
 {
   extern module CentralLB;
 
+  include "CentralLB.h";
+
   initnode void lbinit (void);
 
   group [migratable] RotateLB : CentralLB
