+++ src/xlat-i/xi-Module.C
@@ -58,6 +58,7 @@ void Module::generate()
   "#include \"envelope.h\"\n"
   "#include <memory>\n"
   "#include \"sdag.h\"\n";
+  //"#include \"BaseLB.h\"\n";
 
   if (isTramTarget())
       declstr << "#include \"NDMeshStreamer.h\"\n";
