+++ src/xlat-i/xi-Parameter.h
@@ -34,24 +34,24 @@ class Parameter {
     friend class ParamList;
     void pup(XStr &str);
     void pupArray(XStr &str);
-    void pupRdma(XStr &str, bool genRdma);
+    void pupRdma(XStr &str, int regArray);
     void copyPtr(XStr &str);
     void check();
     void checkPointer(Type *dt);
     void marshallArraySizes(XStr &str,Type *dt);
     void marshallRegArraySizes(XStr &str);
-    void marshallRdmaParameters(XStr &str, bool genRdma);
+    void marshallRdmaParameters(XStr &str, int opts);
     void marshallArrayData(XStr &str);
     void marshallRdmaArrayData(XStr &str);
     void beginUnmarshall(XStr &str);
     void beginUnmarshallArray(XStr &str);
-    void beginUnmarshallRdma(XStr &str, bool genRdma);
+    void beginUnmarshallRdma(XStr &str, int opts);
     void beginUnmarshallSDAGRdma(XStr &str);
     void beginUnmarshallSDAGCall(XStr &str);
-    void beginUnmarshallSDAGCallRdma(XStr &str, bool genRdma);
+    void beginUnmarshallSDAGCallRdma(XStr &str, int opts);
     void unmarshallArrayData(XStr &str);
     void unmarshallRegArrayData(XStr &str);
-    void unmarshallRdmaArrayData(XStr &str, bool genRegArray);
+    void unmarshallRdmaArrayData(XStr &str, int opts);
     void adjustUnmarshalledRdmaPtrsSDAG(XStr &str);
     void unmarshallRegArrayDataSDAG(XStr &str);
     void unmarshallRdmaArrayDataSDAG(XStr &str);
@@ -66,7 +66,7 @@ class Parameter {
     void setConditional(int c);
     void setRdma(bool r);
     void setFirstRdma(bool fr);
-    void print(XStr &str,int withDefaultValues=0,int useConst=1);
+    void print(XStr &str,int withDefaultValues=0,int useConst=1, bool rdmaPost=false);
     void printAddress(XStr &str);
     void printValue(XStr &str);
     int isMessage(void) const;
@@ -105,9 +105,9 @@ class ParamList {
     typedef int (Parameter::*pred_t)(void) const;
     int orEach(pred_t f);
     typedef void (Parameter::*fn_t)(XStr &str);
-    typedef void (Parameter::*rdmafn_t)(XStr &str, bool genRegArray);
+    typedef void (Parameter::*rdmafn_t)(XStr &str, int opts);
     void callEach(fn_t f,XStr &str);
-    void callEach(rdmafn_t f,XStr &str,bool genRegArray);
+    void callEach(rdmafn_t f,XStr &str, int opts);
     void encloseFlag(XStr &str);
     bool manyPointers;
 
@@ -117,7 +117,7 @@ class ParamList {
     ParamList *next;
     ParamList(ParamList *pl);
     ParamList(Parameter *Nparam,ParamList *Nnext=NULL);
-    void print(XStr &str,int withDefaultValues=0,int useConst=1);
+    void print(XStr &str,int withDefaultValues=0,int useConst=1,bool rdmaPost=false);
     void printTypes(XStr &str,int withDefaultValues=0,int useConst=1);
     void printAddress(XStr &str);
     void printValue(XStr &str);
@@ -151,7 +151,8 @@ class ParamList {
     void beginUnmarshallSDAG(XStr &str);
     void beginUnmarshallSDAGCall(XStr &str, bool usesImplBuf);
     void beginRednWrapperUnmarshall(XStr &str, bool isSDAGGen);
-    void unmarshall(XStr &str, int isFirst=1);
+    void beginRdmaWrapperUnmarshall(XStr &str);
+    void unmarshall(XStr &str, int isFirst=1, int opts=0);
     void unmarshallSDAGCall(XStr &str, int isFirst=1);
     void unmarshallAddress(XStr &str, int isFirst=1);
     void pupAllValues(XStr &str);
